Isolate @nestjs/testing dependency
- Extracts `TestAppProvider` to a different file (under `src/__tests__`)
- This change removes the direct reference from `app.provider.ts` (used in production code) to `@nestjs/testing`
