+++ src/app.provider.ts
@@ -1,6 +1,5 @@
 import { INestApplication, VersioningType } from '@nestjs/common';
 import { NestFactory } from '@nestjs/core';
-import { TestingModule } from '@nestjs/testing/testing-module';
 
 function configureVersioning(app: INestApplication) {
   app.enableVersioning({
@@ -14,7 +13,7 @@ function configureCors(app: INestApplication) {
   });
 }
 
-const DEFAULT_CONFIGURATION: ((app: INestApplication) => void)[] = [
+export const DEFAULT_CONFIGURATION: ((app: INestApplication) => void)[] = [
   configureVersioning,
   configureCors,
 ];
@@ -55,30 +54,5 @@ export class DefaultAppProvider extends AppProvider {
 
   protected getApp(module: unknown): Promise<INestApplication> {
     return NestFactory.create(module);
-  }
-}
-
-/**
- * A test {@link AppProvider}
- *
- * This provider provides an application given a {@link TestingModule}
- *
- * If the module provided is not a {@link TestingModule}, an error is thrown
- */
-export class TestAppProvider extends AppProvider {
-  protected readonly configuration: Array<(app: INestApplication) => void> =
-    DEFAULT_CONFIGURATION;
-
-  constructor() {
-    super();
-    if (process.env.NODE_ENV !== 'test') {
-      throw Error('TestAppProvider used outside of a testing environment');
-    }
-  }
-
-  protected getApp(module: unknown): Promise<INestApplication> {
-    if (!(module instanceof TestingModule))
-      return Promise.reject(`Provided module is not a TestingModule`);
-    return Promise.resolve(module.createNestApplication());
   }
 }
